+++ es5.js
@@ -2,6 +2,7 @@ module.exports = {
   extends: ["./index.js"],
 
   rules: {
+    // cancel es6
     "no-var": 0,
   },
 };